Milestone: Set of PoC-Tests Done
With this milestone there are a bunch of tests you could take for
reference implementing your own. The focus is on `WordPressTest` in
order to see, how to test applications which depend on other services
using `DockerComposeContainer`.

For debugging purpose and also for reference added a JUnit 5 version
for the SimpleWebServerTest developed by Baeldung.